refactor: isolate DreamPlace runtime reporting
